refactor: components를 모듈로 변경
