+++ frontend/app/src/components/profile/index.ts
@@ -1,0 +1,4 @@
+export * from './Profile'
+export * from './OtherProfile'
+export * from './ProfileListItem'
+export * from './userActions'
