+++ frontend/app/src/components/utils/index.ts
@@ -1,0 +1,2 @@
+export * from './IconButtonWrap'
+export * from './UserStatus'
